Convert IDNs to ASCII for CSR (fixes IDN support)
The ACME spec requires that identifiers in an order/authz are in the same format they are in the certificate, which is the rules in RFC 5280 section 7, so it is expected that the CSR also have IDNs in ASCII format. Simply relocate the code that does this conversion to the proper place.