Support for device-attest-01 challenge
This commit adds support for the device-attest-01 challenge defined in
https://datatracker.ietf.org/doc/draft-bweeks-acme-device-attest/

This new challenge requires a payload containing an attestation object.
A new Payload field has been added to the Challenge type. If this
payload is set, it will be marshaled to JSON and sent to the CA. By
default, it will be set to "struct{}{}" which will be marshaled as "{}".