Slightly enhance plumbing example
Generating private key or starting the ACME transaction are both expensive in different ways.

A private key is sensitive material though, so maybe it's better to generate it as
late as possible since we don't know how user programs are going to treat it.

There is no technically incorrect way between the two versions.